Change How to Spend It URL (#250)
How to Spend It is moving to FT.com. The stream page for the HTSI brand concept will be the new destination of this link.

Right now, next-stream-page will redirect users without the htsi flag off to https://howtospendit.ft.com. Users with the flag on will be able to access the stream page. After launch of the new HTSI stream page, the flag and redirect to the old site will be removed.